抽出方法
