+++ src/base/conf/app.conf
@@ -7,7 +7,7 @@ EnableDocs = true
 BeegoServerName = JREAM.LU
 EnableGzip = true
 EnableAdmin = true
-ShortUrl = http://goo.lu/
+ShortenDomain = http://goo.lu/
 
 [core]
 author = jream.lu
